+++ src/test/java/game/PlayersTest.java
@@ -1,6 +1,5 @@
 package game;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
@@ -88,21 +87,4 @@ public class PlayersTest {
         Assertions.assertThat(allPlayerNames).containsExactly("aa", "bb");
     }
 
-    @Test
-    void 모든_플레이어의_수익을_계산한다() {
-        //given
-        List<Player> players = List.of(
-                new Player("a", 1000),
-                new Player("b", 1000),
-                new Player("c", 1000),
-                new Player("d", 1000));
-        List<GameResult> gameResults = List.of(
-                GameResult.BLACKJACK, GameResult.WIN, GameResult.DRAW, GameResult.LOSE);
-
-        //when
-        Profits profits = Profits.of(players, gameResults);
-
-        //then
-        assertThat(profits.getProfits()).containsExactlyElementsOf(List.of(1500, 1000, 0, -1000));
-    }
 }
